chore: format

+++ src/lib/MangroveLib.sol
@@ -6,12 +6,16 @@ import {IMangrove, OLKey, Local} from "@mgv/src/IMangrove.sol";
 library MangroveLib {
   using MangroveLib for IMangrove;
 
-  function _minVolume(IMangrove mgv, OLKey memory olKey, uint gasreq) internal view returns (uint256) {
+  function _minVolume(IMangrove mgv, OLKey memory olKey, uint256 gasreq) internal view returns (uint256) {
     Local local = mgv.local(olKey);
     return local.density().multiplyUp(gasreq + local.offer_gasbase());
   }
 
-  function minVolumes(IMangrove mgv, OLKey memory olKey, uint gasreq) internal view returns (uint256 bidVolume, uint256 askVolume) {
+  function minVolumes(IMangrove mgv, OLKey memory olKey, uint256 gasreq)
+    internal
+    view
+    returns (uint256 bidVolume, uint256 askVolume)
+  {
     askVolume = mgv._minVolume(olKey, gasreq);
     bidVolume = mgv._minVolume(olKey.flipped(), gasreq);
   }
